addressed issue when anonymous type on AddKeys. Refactored serialization to avoid multiple #if. tests
